requirements.txt
# App
streamlit
streamlit-folium
folium
Pillow
pandas

# ML
torch
transformers
tqdm

# OCR (experimental)
pytesseract
opencv-python-headless

# Data mining (100% gratuito)
requests
beautifulsoup4
lxml